+++ packages/auto_mappr/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Fix mapping from/to subclasses.
 - Add `tryConvert` method
+- Add `required_inputs` with  `.freezed.dart` and `.drift.dart` by default
 
 ## 1.2.1
 - Document how to use the `convert` method.